# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Multi-Warehouse & Multi-Country Inventory Management System API",
        "description": "This documentation describes all REST endpoints exposed by the Inventory Management backend. All requests (except authentication) are protected with JWT (Bearer) tokens.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Primary API server"
        }
    ],
    "paths": {
...
        "/inventories": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "List all inventories",
                "description": "Get a paginated list of inventories.",
                "operationId": "a817371e777124a5b627194ef89cf950",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InventoryResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Create inventory",
                "description": "Store a newly created inventory record.",
                "operationId": "2cf44ddde1cb9338cbdac1be53bd2336",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inventory created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/inventories/{inventory}": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Get single inventory",
                "description": "Show a specific inventory item.",
                "operationId": "9b5fc42a2b197fbba89318e25814731b",
                "parameters": [
                    {
                        "name": "inventory",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Update inventory",
                "description": "Update a specific inventory item.",
                "operationId": "cf52a722929de7cc9241d5bab15fc5a7",
                "parameters": [
                    {
                        "name": "inventory",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventory updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Delete inventory",
                "description": "Delete a specific inventory item.",
                "operationId": "ad4fae1349b28daa542e118c91daafbf",
                "parameters": [
                    {
                        "name": "inventory",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "422": {
                        "description": "Deletion error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/inventory/global-view": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "Global inventory view",
                "description": "Get a global view of inventory by optional country or warehouse.",
                "operationId": "66c57bfd8fff4354c516d5c87aba35a6",
                "parameters": [
                    {
                        "name": "country_id",
                        "in": "query",
                        "description": "Filter by country ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Filter by warehouse ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Global inventory data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/inventory-transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "List all inventory transactions",
                "operationId": "eabe6be4caa47f2716f49bce4fafed57",
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InventoryTransactionResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Record a new inventory transaction",
                "operationId": "314fa5df72b5586c66e782293bb2d0d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryTransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction successfully recorded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryTransactionResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/inventory-transactions/{inventory_transaction}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get a specific inventory transaction",
                "operationId": "20d6fa21acb07007581409ea71f3325a",
                "parameters": [
                    {
                        "name": "inventory_transaction",
                        "in": "path",
                        "description": "Transaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryTransactionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InventoryRequest": {
                "title": "Inventory Request",
                "description": "Inventory request body data",
                "required": [
                    "product_id",
                    "warehouse_id",
                    "quantity",
                    "min_quantity"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "warehouse_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 150
                    },
                    "min_quantity": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "InventoryTransactionRequest": {
                "title": "Inventory Transaction Request",
                "description": "Inventory transaction request body data",
                "required": [
                    "product_id",
                    "warehouse_id",
                    "quantity",
                    "transaction_type",
                    "date"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "warehouse_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "supplier_id": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 50
                    },
                    "transaction_type": {
                        "type": "string",
                        "enum": [
                            "in",
                            "out"
                        ],
                        "example": "in"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-07-01"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Initial stock",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CountryResource": {
                "title": "Country Resource",
                "description": "Country resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "United States"
                    },
                    "code": {
                        "type": "string",
                        "example": "US"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "InventoryResource": {
                "title": "Inventory Resource",
                "description": "Inventory resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "warehouse_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 150
                    },
                    "min_quantity": {
                        "type": "integer",
                        "example": 20
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "relations": {
                        "properties": {
                            "product": {
                                "$ref": "#/components/schemas/ProductResource"
                            },
                            "warehouse": {
                                "$ref": "#/components/schemas/WarehouseResource"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "InventoryTransactionResource": {
                "title": "Inventory Transaction Resource",
                "description": "Inventory transaction resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "warehouse_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "supplier_id": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 50
                    },
                    "transaction_type": {
                        "type": "string",
                        "example": "in"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-07-01"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_by": {
                        "type": "integer",
                        "example": 1
                    },
                    "notes": {
                        "type": "string",
                        "example": "Initial stock",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProductResource": {
                "title": "Product Resource",
                "description": "Product resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "iPhone 14"
                    },
                    "sku": {
                        "type": "string",
                        "example": "IPH-14-128BLK"
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Apple iPhone 14 128GB Black"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "WarehouseResource": {
                "title": "Warehouse Resource",
                "description": "Warehouse resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "country_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Central Warehouse"
                    },
                    "location": {
                        "type": "string",
                        "example": "New York"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "relations": {
                        "properties": {
                            "country": {
                                "$ref": "#/components/schemas/CountryResource"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}